LayoutTests/platform/mac-monterey-wk2/http/tests/workers/service/openwindow-from-notification-click-expected.txt

This tests that a notificationclick handler can open a window

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Notification.permission is "granted"
FAIL Unexpectedly received the failed-to-show message
